Move <Summary> component to a file, so we can reuse it.
